Remove old Traefik APIs
