+++ tembo-operator/justfile
@@ -18,13 +18,7 @@ install-calico:
 
 install-traefik:
 	kubectl create namespace traefik || true
-	helm upgrade --install --namespace=traefik --version=20.8.0 --values=./testdata/traefik-values.yaml traefik traefik/traefik
-	# https://github.com/traefik/traefik-helm-chart/issues/757#issuecomment-1753995542
-	kubectl apply -f https://raw.githubusercontent.com/traefik/traefik/v3.0.0-beta2/docs/content/reference/dynamic-configuration/kubernetes-crd-definition-v1.yml
-	# Temporary until we upgrade to v3.
-	# First, we have both CRD versions installed, but only using old version.
-	# Then, we upgrade traefik, which cuts over to using the new CRDs, at that point this can be deleted.
-	kubectl apply -f ./testdata/traefik-v3.yaml
+	helm upgrade --install --namespace=traefik --version=28.3.0 --values=./testdata/traefik-values.yaml traefik traefik/traefik
 
 install-operator:
 	just install-cert-manager
