+++ tembo-operator/src/app_service/mod.rs
@@ -1,5 +1,3 @@
 pub mod ingress;
-pub mod ingress_v3;
 pub mod manager;
-pub mod manager_v3;
 pub mod types;
