+++ tembo-operator/src/traefik/middlewares_crd.rs
@@ -1,33 +1,46 @@
 // WARNING: generated by kopium - manual changes will be overwritten
-// kopium command: kopium -A --derive Default middlewares.traefik.containo.us
-// kopium version: 0.14.0
+// kopium command: kopium -A --derive Default middlewares.traefik.io
+// kopium version: 0.20.1
 
-use k8s_openapi::apimachinery::pkg::util::intstr::IntOrString;
-use kube::CustomResource;
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
-use std::collections::BTreeMap;
+#[allow(unused_imports)]
+mod prelude {
+    pub use k8s_openapi::apimachinery::pkg::util::intstr::IntOrString;
+    pub use kube::CustomResource;
+    pub use schemars::JsonSchema;
+    pub use serde::{Deserialize, Serialize};
+    pub use std::collections::BTreeMap;
+}
+use self::prelude::*;
 
 /// MiddlewareSpec defines the desired state of a Middleware.
 #[derive(CustomResource, Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 #[kube(
-    group = "traefik.containo.us",
+    group = "traefik.io",
     version = "v1alpha1",
     kind = "Middleware",
     plural = "middlewares"
 )]
 #[kube(namespaced)]
+#[kube(derive = "Default")]
 pub struct MiddlewareSpec {
-    /// AddPrefix holds the add prefix middleware configuration. This middleware updates the path of a request before forwarding it. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/addprefix/
+    /// AddPrefix holds the add prefix middleware configuration.
+    /// This middleware updates the path of a request before forwarding it.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/addprefix/
     #[serde(default, skip_serializing_if = "Option::is_none", rename = "addPrefix")]
     pub add_prefix: Option<MiddlewareAddPrefix>,
-    /// BasicAuth holds the basic auth middleware configuration. This middleware restricts access to your services to known users. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/basicauth/
+    /// BasicAuth holds the basic auth middleware configuration.
+    /// This middleware restricts access to your services to known users.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/basicauth/
     #[serde(default, skip_serializing_if = "Option::is_none", rename = "basicAuth")]
     pub basic_auth: Option<MiddlewareBasicAuth>,
-    /// Buffering holds the buffering middleware configuration. This middleware retries or limits the size of requests that can be forwarded to backends. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/buffering/#maxrequestbodybytes
+    /// Buffering holds the buffering middleware configuration.
+    /// This middleware retries or limits the size of requests that can be forwarded to backends.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/buffering/#maxrequestbodybytes
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub buffering: Option<MiddlewareBuffering>,
-    /// Chain holds the configuration of the chain middleware. This middleware enables to define reusable combinations of other pieces of middleware. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/chain/
+    /// Chain holds the configuration of the chain middleware.
+    /// This middleware enables to define reusable combinations of other pieces of middleware.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/chain/
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub chain: Option<MiddlewareChain>,
     /// CircuitBreaker holds the circuit breaker configuration.
@@ -37,102 +50,148 @@ pub struct MiddlewareSpec {
         rename = "circuitBreaker"
     )]
     pub circuit_breaker: Option<MiddlewareCircuitBreaker>,
-    /// Compress holds the compress middleware configuration. This middleware compresses responses before sending them to the client, using gzip compression. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/compress/
+    /// Compress holds the compress middleware configuration.
+    /// This middleware compresses responses before sending them to the client, using gzip compression.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/compress/
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub compress: Option<MiddlewareCompress>,
-    /// ContentType holds the content-type middleware configuration. This middleware sets the `Content-Type` header value to the media type detected from the response content, when it is not set by the backend.
+    /// ContentType holds the content-type middleware configuration.
+    /// This middleware exists to enable the correct behavior until at least the default one can be changed in a future version.
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "contentType"
     )]
     pub content_type: Option<MiddlewareContentType>,
-    /// DigestAuth holds the digest auth middleware configuration. This middleware restricts access to your services to known users. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/digestauth/
+    /// DigestAuth holds the digest auth middleware configuration.
+    /// This middleware restricts access to your services to known users.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/digestauth/
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "digestAuth"
     )]
     pub digest_auth: Option<MiddlewareDigestAuth>,
-    /// ErrorPage holds the custom error middleware configuration. This middleware returns a custom page in lieu of the default, according to configured ranges of HTTP Status codes. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/errorpages/
+    /// ErrorPage holds the custom error middleware configuration.
+    /// This middleware returns a custom page in lieu of the default, according to configured ranges of HTTP Status codes.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/errorpages/
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub errors: Option<MiddlewareErrors>,
-    /// ForwardAuth holds the forward auth middleware configuration. This middleware delegates the request authentication to a Service. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/forwardauth/
+    /// ForwardAuth holds the forward auth middleware configuration.
+    /// This middleware delegates the request authentication to a Service.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/forwardauth/
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "forwardAuth"
     )]
     pub forward_auth: Option<MiddlewareForwardAuth>,
-    /// GrpcWeb holds the gRPC web middleware configuration. This middleware converts a gRPC web request to an HTTP/2 gRPC request.
+    /// GrpcWeb holds the gRPC web middleware configuration.
+    /// This middleware converts a gRPC web request to an HTTP/2 gRPC request.
     #[serde(default, skip_serializing_if = "Option::is_none", rename = "grpcWeb")]
     pub grpc_web: Option<MiddlewareGrpcWeb>,
-    /// Headers holds the headers middleware configuration. This middleware manages the requests and responses headers. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/headers/#customrequestheaders
+    /// Headers holds the headers middleware configuration.
+    /// This middleware manages the requests and responses headers.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/headers/#customrequestheaders
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub headers: Option<MiddlewareHeaders>,
-    /// InFlightReq holds the in-flight request middleware configuration. This middleware limits the number of requests being processed and served concurrently. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/inflightreq/
+    /// InFlightReq holds the in-flight request middleware configuration.
+    /// This middleware limits the number of requests being processed and served concurrently.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/inflightreq/
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "inFlightReq"
     )]
     pub in_flight_req: Option<MiddlewareInFlightReq>,
-    /// IPAllowList holds the IP allowlist middleware configuration. This middleware accepts / refuses requests based on the client IP. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ipallowlist/
+    /// IPAllowList holds the IP allowlist middleware configuration.
+    /// This middleware accepts / refuses requests based on the client IP.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ipallowlist/
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "ipAllowList"
     )]
     pub ip_allow_list: Option<MiddlewareIpAllowList>,
-    /// PassTLSClientCert holds the pass TLS client cert middleware configuration. This middleware adds the selected data from the passed client TLS certificate to a header. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/passtlsclientcert/
+    /// Deprecated: please use IPAllowList instead.
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "ipWhiteList"
+    )]
+    pub ip_white_list: Option<MiddlewareIpWhiteList>,
+    /// PassTLSClientCert holds the pass TLS client cert middleware configuration.
+    /// This middleware adds the selected data from the passed client TLS certificate to a header.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/passtlsclientcert/
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "passTLSClientCert"
     )]
-    pub pass_tls_client_cert: Option<MiddlewarePassTLSClientCert>,
-    /// RateLimit holds the rate limit configuration. This middleware ensures that services will receive a fair amount of requests, and allows one to define what fair is. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ratelimit/
+    pub pass_tls_client_cert: Option<MiddlewarePassTlsClientCert>,
+    /// Plugin defines the middleware plugin configuration.
+    /// More info: https://doc.traefik.io/traefik/plugins/
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub plugin: Option<BTreeMap<String, serde_json::Value>>,
+    /// RateLimit holds the rate limit configuration.
+    /// This middleware ensures that services will receive a fair amount of requests, and allows one to define what fair is.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ratelimit/
     #[serde(default, skip_serializing_if = "Option::is_none", rename = "rateLimit")]
     pub rate_limit: Option<MiddlewareRateLimit>,
-    /// RedirectRegex holds the redirect regex middleware configuration. This middleware redirects a request using regex matching and replacement. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/redirectregex/#regex
+    /// RedirectRegex holds the redirect regex middleware configuration.
+    /// This middleware redirects a request using regex matching and replacement.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/redirectregex/#regex
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "redirectRegex"
     )]
     pub redirect_regex: Option<MiddlewareRedirectRegex>,
-    /// RedirectScheme holds the redirect scheme middleware configuration. This middleware redirects requests from a scheme/port to another. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/redirectscheme/
+    /// RedirectScheme holds the redirect scheme middleware configuration.
+    /// This middleware redirects requests from a scheme/port to another.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/redirectscheme/
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "redirectScheme"
     )]
     pub redirect_scheme: Option<MiddlewareRedirectScheme>,
-    /// ReplacePath holds the replace path middleware configuration. This middleware replaces the path of the request URL and store the original path in an X-Replaced-Path header. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/replacepath/
+    /// ReplacePath holds the replace path middleware configuration.
+    /// This middleware replaces the path of the request URL and store the original path in an X-Replaced-Path header.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/replacepath/
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "replacePath"
     )]
     pub replace_path: Option<MiddlewareReplacePath>,
-    /// ReplacePathRegex holds the replace path regex middleware configuration. This middleware replaces the path of a URL using regex matching and replacement. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/replacepathregex/
+    /// ReplacePathRegex holds the replace path regex middleware configuration.
+    /// This middleware replaces the path of a URL using regex matching and replacement.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/replacepathregex/
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "replacePathRegex"
     )]
     pub replace_path_regex: Option<MiddlewareReplacePathRegex>,
-    /// Retry holds the retry middleware configuration. This middleware reissues requests a given number of times to a backend server if that server does not reply. As soon as the server answers, the middleware stops retrying, regardless of the response status. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/retry/
+    /// Retry holds the retry middleware configuration.
+    /// This middleware reissues requests a given number of times to a backend server if that server does not reply.
+    /// As soon as the server answers, the middleware stops retrying, regardless of the response status.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/retry/
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub retry: Option<MiddlewareRetry>,
-    /// StripPrefix holds the strip prefix middleware configuration. This middleware removes the specified prefixes from the URL path. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/stripprefix/
+    /// StripPrefix holds the strip prefix middleware configuration.
+    /// This middleware removes the specified prefixes from the URL path.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/stripprefix/
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "stripPrefix"
     )]
     pub strip_prefix: Option<MiddlewareStripPrefix>,
-    /// StripPrefixRegex holds the strip prefix regex middleware configuration. This middleware removes the matching prefixes from the URL path. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/stripprefixregex/
+    /// StripPrefixRegex holds the strip prefix regex middleware configuration.
+    /// This middleware removes the matching prefixes from the URL path.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/stripprefixregex/
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
@@ -141,28 +200,36 @@ pub struct MiddlewareSpec {
     pub strip_prefix_regex: Option<MiddlewareStripPrefixRegex>,
 }
 
-/// AddPrefix holds the add prefix middleware configuration. This middleware updates the path of a request before forwarding it. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/addprefix/
+/// AddPrefix holds the add prefix middleware configuration.
+/// This middleware updates the path of a request before forwarding it.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/addprefix/
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareAddPrefix {
-    /// Prefix is the string to add before the current path in the requested URL. It should include a leading slash (/).
+    /// Prefix is the string to add before the current path in the requested URL.
+    /// It should include a leading slash (/).
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub prefix: Option<String>,
 }
 
-/// BasicAuth holds the basic auth middleware configuration. This middleware restricts access to your services to known users. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/basicauth/
+/// BasicAuth holds the basic auth middleware configuration.
+/// This middleware restricts access to your services to known users.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/basicauth/
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareBasicAuth {
-    /// HeaderField defines a header field to store the authenticated user. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/basicauth/#headerfield
+    /// HeaderField defines a header field to store the authenticated user.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/basicauth/#headerfield
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "headerField"
     )]
     pub header_field: Option<String>,
-    /// Realm allows the protected resources on a server to be partitioned into a set of protection spaces, each with its own authentication scheme. Default: traefik.
+    /// Realm allows the protected resources on a server to be partitioned into a set of protection spaces, each with its own authentication scheme.
+    /// Default: traefik.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub realm: Option<String>,
-    /// RemoveHeader sets the removeHeader option to true to remove the authorization header before forwarding the request to your service. Default: false.
+    /// RemoveHeader sets the removeHeader option to true to remove the authorization header before forwarding the request to your service.
+    /// Default: false.
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
@@ -174,38 +241,48 @@ pub struct MiddlewareBasicAuth {
     pub secret: Option<String>,
 }
 
-/// Buffering holds the buffering middleware configuration. This middleware retries or limits the size of requests that can be forwarded to backends. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/buffering/#maxrequestbodybytes
+/// Buffering holds the buffering middleware configuration.
+/// This middleware retries or limits the size of requests that can be forwarded to backends.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/buffering/#maxrequestbodybytes
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareBuffering {
-    /// MaxRequestBodyBytes defines the maximum allowed body size for the request (in bytes). If the request exceeds the allowed size, it is not forwarded to the service, and the client gets a 413 (Request Entity Too Large) response. Default: 0 (no maximum).
+    /// MaxRequestBodyBytes defines the maximum allowed body size for the request (in bytes).
+    /// If the request exceeds the allowed size, it is not forwarded to the service, and the client gets a 413 (Request Entity Too Large) response.
+    /// Default: 0 (no maximum).
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "maxRequestBodyBytes"
     )]
     pub max_request_body_bytes: Option<i64>,
-    /// MaxResponseBodyBytes defines the maximum allowed response size from the service (in bytes). If the response exceeds the allowed size, it is not forwarded to the client. The client gets a 500 (Internal Server Error) response instead. Default: 0 (no maximum).
+    /// MaxResponseBodyBytes defines the maximum allowed response size from the service (in bytes).
+    /// If the response exceeds the allowed size, it is not forwarded to the client. The client gets a 500 (Internal Server Error) response instead.
+    /// Default: 0 (no maximum).
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "maxResponseBodyBytes"
     )]
     pub max_response_body_bytes: Option<i64>,
-    /// MemRequestBodyBytes defines the threshold (in bytes) from which the request will be buffered on disk instead of in memory. Default: 1048576 (1Mi).
+    /// MemRequestBodyBytes defines the threshold (in bytes) from which the request will be buffered on disk instead of in memory.
+    /// Default: 1048576 (1Mi).
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "memRequestBodyBytes"
     )]
     pub mem_request_body_bytes: Option<i64>,
-    /// MemResponseBodyBytes defines the threshold (in bytes) from which the response will be buffered on disk instead of in memory. Default: 1048576 (1Mi).
+    /// MemResponseBodyBytes defines the threshold (in bytes) from which the response will be buffered on disk instead of in memory.
+    /// Default: 1048576 (1Mi).
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "memResponseBodyBytes"
     )]
     pub mem_response_body_bytes: Option<i64>,
-    /// RetryExpression defines the retry conditions. It is a logical combination of functions with operators AND (&&) and OR (||). More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/buffering/#retryexpression
+    /// RetryExpression defines the retry conditions.
+    /// It is a logical combination of functions with operators AND (&&) and OR (||).
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/buffering/#retryexpression
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
@@ -214,7 +291,9 @@ pub struct MiddlewareBuffering {
     pub retry_expression: Option<String>,
 }
 
-/// Chain holds the configuration of the chain middleware. This middleware enables to define reusable combinations of other pieces of middleware. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/chain/
+/// Chain holds the configuration of the chain middleware.
+/// This middleware enables to define reusable combinations of other pieces of middleware.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/chain/
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareChain {
     /// Middlewares is the list of MiddlewareRef which composes the chain.
@@ -261,17 +340,28 @@ pub struct MiddlewareCircuitBreaker {
     pub recovery_duration: Option<IntOrString>,
 }
 
-/// Compress holds the compress middleware configuration. This middleware compresses responses before sending them to the client, using gzip compression. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/compress/
+/// Compress holds the compress middleware configuration.
+/// This middleware compresses responses before sending them to the client, using gzip compression.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/compress/
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareCompress {
-    /// ExcludedContentTypes defines the list of content types to compare the Content-Type header of the incoming requests and responses before compressing. `application/grpc` is always excluded.
+    /// ExcludedContentTypes defines the list of content types to compare the Content-Type header of the incoming requests and responses before compressing.
+    /// `application/grpc` is always excluded.
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "excludedContentTypes"
     )]
     pub excluded_content_types: Option<Vec<String>>,
-    /// MinResponseBodyBytes defines the minimum amount of bytes a response body must have to be compressed. Default: 1024.
+    /// IncludedContentTypes defines the list of content types to compare the Content-Type header of the responses before compressing.
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "includedContentTypes"
+    )]
+    pub included_content_types: Option<Vec<String>>,
+    /// MinResponseBodyBytes defines the minimum amount of bytes a response body must have to be compressed.
+    /// Default: 1024.
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
@@ -280,21 +370,36 @@ pub struct MiddlewareCompress {
     pub min_response_body_bytes: Option<i64>,
 }
 
-/// ContentType holds the content-type middleware configuration. This middleware sets the `Content-Type` header value to the media type detected from the response content, when it is not set by the backend.
+/// ContentType holds the content-type middleware configuration.
+/// This middleware exists to enable the correct behavior until at least the default one can be changed in a future version.
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
-pub struct MiddlewareContentType {}
+pub struct MiddlewareContentType {
+    /// AutoDetect specifies whether to let the `Content-Type` header, if it has not been set by the backend,
+    /// be automatically set to a value derived from the contents of the response.
+    /// Deprecated: AutoDetect option is deprecated, Content-Type middleware is only meant to be used to enable the content-type detection, please remove any usage of this option.
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "autoDetect"
+    )]
+    pub auto_detect: Option<bool>,
+}
 
-/// DigestAuth holds the digest auth middleware configuration. This middleware restricts access to your services to known users. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/digestauth/
+/// DigestAuth holds the digest auth middleware configuration.
+/// This middleware restricts access to your services to known users.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/digestauth/
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareDigestAuth {
-    /// HeaderField defines a header field to store the authenticated user. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/basicauth/#headerfield
+    /// HeaderField defines a header field to store the authenticated user.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/basicauth/#headerfield
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "headerField"
     )]
     pub header_field: Option<String>,
-    /// Realm allows the protected resources on a server to be partitioned into a set of protection spaces, each with its own authentication scheme. Default: traefik.
+    /// Realm allows the protected resources on a server to be partitioned into a set of protection spaces, each with its own authentication scheme.
+    /// Default: traefik.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub realm: Option<String>,
     /// RemoveHeader defines whether to remove the authorization header before forwarding the request to the backend.
@@ -309,39 +414,57 @@ pub struct MiddlewareDigestAuth {
     pub secret: Option<String>,
 }
 
-/// ErrorPage holds the custom error middleware configuration. This middleware returns a custom page in lieu of the default, according to configured ranges of HTTP Status codes. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/errorpages/
+/// ErrorPage holds the custom error middleware configuration.
+/// This middleware returns a custom page in lieu of the default, according to configured ranges of HTTP Status codes.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/errorpages/
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareErrors {
-    /// Query defines the URL for the error page (hosted by service). The {status} variable can be used in order to insert the status code in the URL.
+    /// Query defines the URL for the error page (hosted by service).
+    /// The {status} variable can be used in order to insert the status code in the URL.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub query: Option<String>,
-    /// Service defines the reference to a Kubernetes Service that will serve the error page. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/errorpages/#service
+    /// Service defines the reference to a Kubernetes Service that will serve the error page.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/errorpages/#service
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub service: Option<MiddlewareErrorsService>,
-    /// Status defines which status or range of statuses should result in an error page. It can be either a status code as a number (500), as multiple comma-separated numbers (500,502), as ranges by separating two codes with a dash (500-599), or a combination of the two (404,418,500-599).
+    /// Status defines which status or range of statuses should result in an error page.
+    /// It can be either a status code as a number (500),
+    /// as multiple comma-separated numbers (500,502),
+    /// as ranges by separating two codes with a dash (500-599),
+    /// or a combination of the two (404,418,500-599).
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub status: Option<Vec<String>>,
 }
 
-/// Service defines the reference to a Kubernetes Service that will serve the error page. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/errorpages/#service
+/// Service defines the reference to a Kubernetes Service that will serve the error page.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/errorpages/#service
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareErrorsService {
     /// Kind defines the kind of the Service.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub kind: Option<MiddlewareErrorsServiceKind>,
-    /// Name defines the name of the referenced Kubernetes Service or TraefikService. The differentiation between the two is specified in the Kind field.
+    /// Name defines the name of the referenced Kubernetes Service or TraefikService.
+    /// The differentiation between the two is specified in the Kind field.
     pub name: String,
     /// Namespace defines the namespace of the referenced Kubernetes Service or TraefikService.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub namespace: Option<String>,
-    /// PassHostHeader defines whether the client Host header is forwarded to the upstream Kubernetes Service. By default, passHostHeader is true.
+    /// NativeLB controls, when creating the load-balancer,
+    /// whether the LB's children are directly the pods IPs or if the only child is the Kubernetes Service clusterIP.
+    /// The Kubernetes Service itself does load-balance to the pods.
+    /// By default, NativeLB is false.
+    #[serde(default, skip_serializing_if = "Option::is_none", rename = "nativeLB")]
+    pub native_lb: Option<bool>,
+    /// PassHostHeader defines whether the client Host header is forwarded to the upstream Kubernetes Service.
+    /// By default, passHostHeader is true.
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "passHostHeader"
     )]
     pub pass_host_header: Option<bool>,
-    /// Port defines the port of a Kubernetes Service. This can be a reference to a named port.
+    /// Port defines the port of a Kubernetes Service.
+    /// This can be a reference to a named port.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub port: Option<IntOrString>,
     /// ResponseForwarding defines how Traefik forwards the response from the upstream Kubernetes Service to the client.
@@ -351,28 +474,35 @@ pub struct MiddlewareErrorsService {
         rename = "responseForwarding"
     )]
     pub response_forwarding: Option<MiddlewareErrorsServiceResponseForwarding>,
-    /// Scheme defines the scheme to use for the request to the upstream Kubernetes Service. It defaults to https when Kubernetes Service port is 443, http otherwise.
+    /// Scheme defines the scheme to use for the request to the upstream Kubernetes Service.
+    /// It defaults to https when Kubernetes Service port is 443, http otherwise.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub scheme: Option<String>,
-    /// ServersTransport defines the name of ServersTransport resource to use. It allows to configure the transport between Traefik and your servers. Can only be used on a Kubernetes Service.
+    /// ServersTransport defines the name of ServersTransport resource to use.
+    /// It allows to configure the transport between Traefik and your servers.
+    /// Can only be used on a Kubernetes Service.
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "serversTransport"
     )]
     pub servers_transport: Option<String>,
-    /// Sticky defines the sticky sessions configuration. More info: https://doc.traefik.io/traefik/v3.0/routing/services/#sticky-sessions
+    /// Sticky defines the sticky sessions configuration.
+    /// More info: https://doc.traefik.io/traefik/v3.0/routing/services/#sticky-sessions
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub sticky: Option<MiddlewareErrorsServiceSticky>,
-    /// Strategy defines the load balancing strategy between the servers. RoundRobin is the only supported value at the moment.
+    /// Strategy defines the load balancing strategy between the servers.
+    /// RoundRobin is the only supported value at the moment.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub strategy: Option<String>,
-    /// Weight defines the weight and should only be specified when Name references a TraefikService object (and to be precise, one that embeds a Weighted Round Robin).
+    /// Weight defines the weight and should only be specified when Name references a TraefikService object
+    /// (and to be precise, one that embeds a Weighted Round Robin).
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub weight: Option<i64>,
 }
 
-/// Service defines the reference to a Kubernetes Service that will serve the error page. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/errorpages/#service
+/// Service defines the reference to a Kubernetes Service that will serve the error page.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/errorpages/#service
 #[derive(Serialize, Deserialize, Clone, Debug, JsonSchema)]
 pub enum MiddlewareErrorsServiceKind {
     Service,
@@ -382,7 +512,11 @@ pub enum MiddlewareErrorsServiceKind {
 /// ResponseForwarding defines how Traefik forwards the response from the upstream Kubernetes Service to the client.
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareErrorsServiceResponseForwarding {
-    /// FlushInterval defines the interval, in milliseconds, in between flushes to the client while copying the response body. A negative value means to flush immediately after each write to the client. This configuration is ignored when ReverseProxy recognizes a response as a streaming response; for such responses, writes are flushed to the client immediately. Default: 100ms
+    /// FlushInterval defines the interval, in milliseconds, in between flushes to the client while copying the response body.
+    /// A negative value means to flush immediately after each write to the client.
+    /// This configuration is ignored when ReverseProxy recognizes a response as a streaming response;
+    /// for such responses, writes are flushed to the client immediately.
+    /// Default: 100ms
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
@@ -391,7 +525,8 @@ pub struct MiddlewareErrorsServiceResponseForwarding {
     pub flush_interval: Option<String>,
 }
 
-/// Sticky defines the sticky sessions configuration. More info: https://doc.traefik.io/traefik/v3.0/routing/services/#sticky-sessions
+/// Sticky defines the sticky sessions configuration.
+/// More info: https://doc.traefik.io/traefik/v3.0/routing/services/#sticky-sessions
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareErrorsServiceSticky {
     /// Cookie defines the sticky cookie configuration.
@@ -405,10 +540,16 @@ pub struct MiddlewareErrorsServiceStickyCookie {
     /// HTTPOnly defines whether the cookie can be accessed by client-side APIs, such as JavaScript.
     #[serde(default, skip_serializing_if = "Option::is_none", rename = "httpOnly")]
     pub http_only: Option<bool>,
+    /// MaxAge indicates the number of seconds until the cookie expires.
+    /// When set to a negative number, the cookie expires immediately.
+    /// When set to zero, the cookie never expires.
+    #[serde(default, skip_serializing_if = "Option::is_none", rename = "maxAge")]
+    pub max_age: Option<i64>,
     /// Name defines the Cookie name.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub name: Option<String>,
-    /// SameSite defines the same site policy. More info: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie/SameSite
+    /// SameSite defines the same site policy.
+    /// More info: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie/SameSite
     #[serde(default, skip_serializing_if = "Option::is_none", rename = "sameSite")]
     pub same_site: Option<String>,
     /// Secure defines whether the cookie can only be transmitted over an encrypted connection (i.e. HTTPS).
@@ -416,13 +557,23 @@ pub struct MiddlewareErrorsServiceStickyCookie {
     pub secure: Option<bool>,
 }
 
-/// ForwardAuth holds the forward auth middleware configuration. This middleware delegates the request authentication to a Service. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/forwardauth/
+/// ForwardAuth holds the forward auth middleware configuration.
+/// This middleware delegates the request authentication to a Service.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/forwardauth/
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareForwardAuth {
+    /// AddAuthCookiesToResponse defines the list of cookies to copy from the authentication server response to the response.
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "addAuthCookiesToResponse"
+    )]
+    pub add_auth_cookies_to_response: Option<Vec<String>>,
     /// Address defines the authentication server address.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub address: Option<String>,
-    /// AuthRequestHeaders defines the list of the headers to copy from the request to the authentication server. If not set or empty then all request headers are passed.
+    /// AuthRequestHeaders defines the list of the headers to copy from the request to the authentication server.
+    /// If not set or empty then all request headers are passed.
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
@@ -436,7 +587,8 @@ pub struct MiddlewareForwardAuth {
         rename = "authResponseHeaders"
     )]
     pub auth_response_headers: Option<Vec<String>>,
-    /// AuthResponseHeadersRegex defines the regex to match headers to copy from the authentication server response and set on forwarded request, after stripping all headers that match the regex. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/forwardauth/#authresponseheadersregex
+    /// AuthResponseHeadersRegex defines the regex to match headers to copy from the authentication server response and set on forwarded request, after stripping all headers that match the regex.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/forwardauth/#authresponseheadersregex
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
@@ -458,10 +610,19 @@ pub struct MiddlewareForwardAuth {
 /// TLS defines the configuration used to secure the connection to the authentication server.
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareForwardAuthTls {
-    /// CASecret is the name of the referenced Kubernetes Secret containing the CA to validate the server certificate. The CA certificate is extracted from key `tls.ca` or `ca.crt`.
+    /// Deprecated: TLS client authentication is a server side option (see https://github.com/golang/go/blob/740a490f71d026bb7d2d13cb8fa2d6d6e0572b70/src/crypto/tls/common.go#L634).
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "caOptional"
+    )]
+    pub ca_optional: Option<bool>,
+    /// CASecret is the name of the referenced Kubernetes Secret containing the CA to validate the server certificate.
+    /// The CA certificate is extracted from key `tls.ca` or `ca.crt`.
     #[serde(default, skip_serializing_if = "Option::is_none", rename = "caSecret")]
     pub ca_secret: Option<String>,
-    /// CertSecret is the name of the referenced Kubernetes Secret containing the client certificate. The client certificate is extracted from the keys `tls.crt` and `tls.key`.
+    /// CertSecret is the name of the referenced Kubernetes Secret containing the client certificate.
+    /// The client certificate is extracted from the keys `tls.crt` and `tls.key`.
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
@@ -477,10 +638,12 @@ pub struct MiddlewareForwardAuthTls {
     pub insecure_skip_verify: Option<bool>,
 }
 
-/// GrpcWeb holds the gRPC web middleware configuration. This middleware converts a gRPC web request to an HTTP/2 gRPC request.
+/// GrpcWeb holds the gRPC web middleware configuration.
+/// This middleware converts a gRPC web request to an HTTP/2 gRPC request.
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareGrpcWeb {
-    /// AllowOrigins is a list of allowable origins. Can also be a wildcard origin "*".
+    /// AllowOrigins is a list of allowable origins.
+    /// Can also be a wildcard origin "*".
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
@@ -489,7 +652,9 @@ pub struct MiddlewareGrpcWeb {
     pub allow_origins: Option<Vec<String>>,
 }
 
-/// Headers holds the headers middleware configuration. This middleware manages the requests and responses headers. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/headers/#customrequestheaders
+/// Headers holds the headers middleware configuration.
+/// This middleware manages the requests and responses headers.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/headers/#customrequestheaders
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareHeaders {
     /// AccessControlAllowCredentials defines whether the request can include user credentials.
@@ -576,14 +741,16 @@ pub struct MiddlewareHeaders {
         rename = "contentTypeNosniff"
     )]
     pub content_type_nosniff: Option<bool>,
-    /// CustomBrowserXSSValue defines the X-XSS-Protection header value. This overrides the BrowserXssFilter option.
+    /// CustomBrowserXSSValue defines the X-XSS-Protection header value.
+    /// This overrides the BrowserXssFilter option.
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "customBrowserXSSValue"
     )]
     pub custom_browser_xss_value: Option<String>,
-    /// CustomFrameOptionsValue defines the X-Frame-Options header value. This overrides the FrameDeny option.
+    /// CustomFrameOptionsValue defines the X-Frame-Options header value.
+    /// This overrides the FrameDeny option.
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
@@ -604,6 +771,13 @@ pub struct MiddlewareHeaders {
         rename = "customResponseHeaders"
     )]
     pub custom_response_headers: Option<BTreeMap<String, String>>,
+    /// Deprecated: FeaturePolicy option is deprecated, please use PermissionsPolicy instead.
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "featurePolicy"
+    )]
+    pub feature_policy: Option<String>,
     /// ForceSTSHeader defines whether to add the STS header even when the connection is HTTP.
     #[serde(
         default,
@@ -621,14 +795,18 @@ pub struct MiddlewareHeaders {
         rename = "hostsProxyHeaders"
     )]
     pub hosts_proxy_headers: Option<Vec<String>>,
-    /// IsDevelopment defines whether to mitigate the unwanted effects of the AllowedHosts, SSL, and STS options when developing. Usually testing takes place using HTTP, not HTTPS, and on localhost, not your production domain. If you would like your development environment to mimic production with complete Host blocking, SSL redirects, and STS headers, leave this as false.
+    /// IsDevelopment defines whether to mitigate the unwanted effects of the AllowedHosts, SSL, and STS options when developing.
+    /// Usually testing takes place using HTTP, not HTTPS, and on localhost, not your production domain.
+    /// If you would like your development environment to mimic production with complete Host blocking, SSL redirects,
+    /// and STS headers, leave this as false.
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "isDevelopment"
     )]
     pub is_development: Option<bool>,
-    /// PermissionsPolicy defines the Permissions-Policy header value. This allows sites to control browser features.
+    /// PermissionsPolicy defines the Permissions-Policy header value.
+    /// This allows sites to control browser features.
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
@@ -638,20 +816,46 @@ pub struct MiddlewareHeaders {
     /// PublicKey is the public key that implements HPKP to prevent MITM attacks with forged certificates.
     #[serde(default, skip_serializing_if = "Option::is_none", rename = "publicKey")]
     pub public_key: Option<String>,
-    /// ReferrerPolicy defines the Referrer-Policy header value. This allows sites to control whether browsers forward the Referer header to other sites.
+    /// ReferrerPolicy defines the Referrer-Policy header value.
+    /// This allows sites to control whether browsers forward the Referer header to other sites.
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "referrerPolicy"
     )]
     pub referrer_policy: Option<String>,
-    /// SSLProxyHeaders defines the header keys with associated values that would indicate a valid HTTPS request. It can be useful when using other proxies (example: "X-Forwarded-Proto": "https").
+    /// Deprecated: SSLForceHost option is deprecated, please use RedirectRegex instead.
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "sslForceHost"
+    )]
+    pub ssl_force_host: Option<bool>,
+    /// Deprecated: SSLHost option is deprecated, please use RedirectRegex instead.
+    #[serde(default, skip_serializing_if = "Option::is_none", rename = "sslHost")]
+    pub ssl_host: Option<String>,
+    /// SSLProxyHeaders defines the header keys with associated values that would indicate a valid HTTPS request.
+    /// It can be useful when using other proxies (example: "X-Forwarded-Proto": "https").
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "sslProxyHeaders"
     )]
     pub ssl_proxy_headers: Option<BTreeMap<String, String>>,
+    /// Deprecated: SSLRedirect option is deprecated, please use EntryPoint redirection or RedirectScheme instead.
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "sslRedirect"
+    )]
+    pub ssl_redirect: Option<bool>,
+    /// Deprecated: SSLTemporaryRedirect option is deprecated, please use EntryPoint redirection or RedirectScheme instead.
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "sslTemporaryRedirect"
+    )]
+    pub ssl_temporary_redirect: Option<bool>,
     /// STSIncludeSubdomains defines whether the includeSubDomains directive is appended to the Strict-Transport-Security header.
     #[serde(
         default,
@@ -666,7 +870,8 @@ pub struct MiddlewareHeaders {
         rename = "stsPreload"
     )]
     pub sts_preload: Option<bool>,
-    /// STSSeconds defines the max-age of the Strict-Transport-Security header. If set to 0, the header is not set.
+    /// STSSeconds defines the max-age of the Strict-Transport-Security header.
+    /// If set to 0, the header is not set.
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
@@ -675,13 +880,19 @@ pub struct MiddlewareHeaders {
     pub sts_seconds: Option<i64>,
 }
 
-/// InFlightReq holds the in-flight request middleware configuration. This middleware limits the number of requests being processed and served concurrently. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/inflightreq/
+/// InFlightReq holds the in-flight request middleware configuration.
+/// This middleware limits the number of requests being processed and served concurrently.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/inflightreq/
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareInFlightReq {
-    /// Amount defines the maximum amount of allowed simultaneous in-flight request. The middleware responds with HTTP 429 Too Many Requests if there are already amount requests in progress (based on the same sourceCriterion strategy).
+    /// Amount defines the maximum amount of allowed simultaneous in-flight request.
+    /// The middleware responds with HTTP 429 Too Many Requests if there are already amount requests in progress (based on the same sourceCriterion strategy).
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub amount: Option<i64>,
-    /// SourceCriterion defines what criterion is used to group requests as originating from a common source. If several strategies are defined at the same time, an error will be raised. If none are set, the default is to use the requestHost. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/inflightreq/#sourcecriterion
+    /// SourceCriterion defines what criterion is used to group requests as originating from a common source.
+    /// If several strategies are defined at the same time, an error will be raised.
+    /// If none are set, the default is to use the requestHost.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/inflightreq/#sourcecriterion
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
@@ -690,10 +901,14 @@ pub struct MiddlewareInFlightReq {
     pub source_criterion: Option<MiddlewareInFlightReqSourceCriterion>,
 }
 
-/// SourceCriterion defines what criterion is used to group requests as originating from a common source. If several strategies are defined at the same time, an error will be raised. If none are set, the default is to use the requestHost. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/inflightreq/#sourcecriterion
+/// SourceCriterion defines what criterion is used to group requests as originating from a common source.
+/// If several strategies are defined at the same time, an error will be raised.
+/// If none are set, the default is to use the requestHost.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/inflightreq/#sourcecriterion
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareInFlightReqSourceCriterion {
-    /// IPStrategy holds the IP strategy configuration used by Traefik to determine the client IP. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ipallowlist/#ipstrategy
+    /// IPStrategy holds the IP strategy configuration used by Traefik to determine the client IP.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ipallowlist/#ipstrategy
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
@@ -716,7 +931,8 @@ pub struct MiddlewareInFlightReqSourceCriterion {
     pub request_host: Option<bool>,
 }
 
-/// IPStrategy holds the IP strategy configuration used by Traefik to determine the client IP. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ipallowlist/#ipstrategy
+/// IPStrategy holds the IP strategy configuration used by Traefik to determine the client IP.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ipallowlist/#ipstrategy
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareInFlightReqSourceCriterionIpStrategy {
     /// Depth tells Traefik to use the X-Forwarded-For header and take the IP located at the depth position (starting from the right).
@@ -731,16 +947,27 @@ pub struct MiddlewareInFlightReqSourceCriterionIpStrategy {
     pub excluded_i_ps: Option<Vec<String>>,
 }
 
-/// IPAllowList holds the IP allowlist middleware configuration. This middleware accepts / refuses requests based on the client IP. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ipallowlist/
+/// IPAllowList holds the IP allowlist middleware configuration.
+/// This middleware accepts / refuses requests based on the client IP.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ipallowlist/
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareIpAllowList {
-    /// IPStrategy holds the IP strategy configuration used by Traefik to determine the client IP. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ipallowlist/#ipstrategy
+    /// IPStrategy holds the IP strategy configuration used by Traefik to determine the client IP.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ipallowlist/#ipstrategy
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
         rename = "ipStrategy"
     )]
     pub ip_strategy: Option<MiddlewareIpAllowListIpStrategy>,
+    /// RejectStatusCode defines the HTTP status code used for refused requests.
+    /// If not set, the default is 403 (Forbidden).
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "rejectStatusCode"
+    )]
+    pub reject_status_code: Option<i64>,
     /// SourceRange defines the set of allowed IPs (or ranges of allowed IPs by using CIDR notation).
     #[serde(
         default,
@@ -750,7 +977,8 @@ pub struct MiddlewareIpAllowList {
     pub source_range: Option<Vec<String>>,
 }
 
-/// IPStrategy holds the IP strategy configuration used by Traefik to determine the client IP. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ipallowlist/#ipstrategy
+/// IPStrategy holds the IP strategy configuration used by Traefik to determine the client IP.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ipallowlist/#ipstrategy
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareIpAllowListIpStrategy {
     /// Depth tells Traefik to use the X-Forwarded-For header and take the IP located at the depth position (starting from the right).
@@ -765,12 +993,50 @@ pub struct MiddlewareIpAllowListIpStrategy {
     pub excluded_i_ps: Option<Vec<String>>,
 }
 
-/// PassTLSClientCert holds the pass TLS client cert middleware configuration. This middleware adds the selected data from the passed client TLS certificate to a header. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/passtlsclientcert/
+/// Deprecated: please use IPAllowList instead.
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
-pub struct MiddlewarePassTLSClientCert {
+pub struct MiddlewareIpWhiteList {
+    /// IPStrategy holds the IP strategy configuration used by Traefik to determine the client IP.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ipallowlist/#ipstrategy
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "ipStrategy"
+    )]
+    pub ip_strategy: Option<MiddlewareIpWhiteListIpStrategy>,
+    /// SourceRange defines the set of allowed IPs (or ranges of allowed IPs by using CIDR notation).
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "sourceRange"
+    )]
+    pub source_range: Option<Vec<String>>,
+}
+
+/// IPStrategy holds the IP strategy configuration used by Traefik to determine the client IP.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ipallowlist/#ipstrategy
+#[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
+pub struct MiddlewareIpWhiteListIpStrategy {
+    /// Depth tells Traefik to use the X-Forwarded-For header and take the IP located at the depth position (starting from the right).
+    #[serde(default, skip_serializing_if = "Option::is_none")]
+    pub depth: Option<i64>,
+    /// ExcludedIPs configures Traefik to scan the X-Forwarded-For header and select the first IP not in the list.
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "excludedIPs"
+    )]
+    pub excluded_i_ps: Option<Vec<String>>,
+}
+
+/// PassTLSClientCert holds the pass TLS client cert middleware configuration.
+/// This middleware adds the selected data from the passed client TLS certificate to a header.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/passtlsclientcert/
+#[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
+pub struct MiddlewarePassTlsClientCert {
     /// Info selects the specific client certificate details you want to add to the X-Forwarded-Tls-Client-Cert-Info header.
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub info: Option<MiddlewarePassTLSClientCertInfo>,
+    pub info: Option<MiddlewarePassTlsClientCertInfo>,
     /// PEM sets the X-Forwarded-Tls-Client-Cert header with the certificate.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub pem: Option<bool>,
@@ -778,10 +1044,10 @@ pub struct MiddlewarePassTLSClientCert {
 
 /// Info selects the specific client certificate details you want to add to the X-Forwarded-Tls-Client-Cert-Info header.
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
-pub struct MiddlewarePassTLSClientCertInfo {
+pub struct MiddlewarePassTlsClientCertInfo {
     /// Issuer defines the client certificate issuer details to add to the X-Forwarded-Tls-Client-Cert-Info header.
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub issuer: Option<MiddlewarePassTLSClientCertInfoIssuer>,
+    pub issuer: Option<MiddlewarePassTlsClientCertInfoIssuer>,
     /// NotAfter defines whether to add the Not After information from the Validity part.
     #[serde(default, skip_serializing_if = "Option::is_none", rename = "notAfter")]
     pub not_after: Option<bool>,
@@ -800,12 +1066,12 @@ pub struct MiddlewarePassTLSClientCertInfo {
     pub serial_number: Option<bool>,
     /// Subject defines the client certificate subject details to add to the X-Forwarded-Tls-Client-Cert-Info header.
     #[serde(default, skip_serializing_if = "Option::is_none")]
-    pub subject: Option<MiddlewarePassTLSClientCertInfoSubject>,
+    pub subject: Option<MiddlewarePassTlsClientCertInfoSubject>,
 }
 
 /// Issuer defines the client certificate issuer details to add to the X-Forwarded-Tls-Client-Cert-Info header.
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
-pub struct MiddlewarePassTLSClientCertInfoIssuer {
+pub struct MiddlewarePassTlsClientCertInfoIssuer {
     /// CommonName defines whether to add the organizationalUnit information into the issuer.
     #[serde(
         default,
@@ -843,7 +1109,7 @@ pub struct MiddlewarePassTLSClientCertInfoIssuer {
 
 /// Subject defines the client certificate subject details to add to the X-Forwarded-Tls-Client-Cert-Info header.
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
-pub struct MiddlewarePassTLSClientCertInfoSubject {
+pub struct MiddlewarePassTlsClientCertInfoSubject {
     /// CommonName defines whether to add the organizationalUnit information into the subject.
     #[serde(
         default,
@@ -886,19 +1152,28 @@ pub struct MiddlewarePassTLSClientCertInfoSubject {
     pub serial_number: Option<bool>,
 }
 
-/// RateLimit holds the rate limit configuration. This middleware ensures that services will receive a fair amount of requests, and allows one to define what fair is. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ratelimit/
+/// RateLimit holds the rate limit configuration.
+/// This middleware ensures that services will receive a fair amount of requests, and allows one to define what fair is.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ratelimit/
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareRateLimit {
-    /// Average is the maximum rate, by default in requests/s, allowed for the given source. It defaults to 0, which means no rate limiting. The rate is actually defined by dividing Average by Period. So for a rate below 1req/s, one needs to define a Period larger than a second.
+    /// Average is the maximum rate, by default in requests/s, allowed for the given source.
+    /// It defaults to 0, which means no rate limiting.
+    /// The rate is actually defined by dividing Average by Period. So for a rate below 1req/s,
+    /// one needs to define a Period larger than a second.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub average: Option<i64>,
-    /// Burst is the maximum number of requests allowed to arrive in the same arbitrarily small period of time. It defaults to 1.
+    /// Burst is the maximum number of requests allowed to arrive in the same arbitrarily small period of time.
+    /// It defaults to 1.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub burst: Option<i64>,
-    /// Period, in combination with Average, defines the actual maximum rate, such as: r = Average / Period. It defaults to a second.
+    /// Period, in combination with Average, defines the actual maximum rate, such as:
+    /// r = Average / Period. It defaults to a second.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub period: Option<IntOrString>,
-    /// SourceCriterion defines what criterion is used to group requests as originating from a common source. If several strategies are defined at the same time, an error will be raised. If none are set, the default is to use the request's remote address field (as an ipStrategy).
+    /// SourceCriterion defines what criterion is used to group requests as originating from a common source.
+    /// If several strategies are defined at the same time, an error will be raised.
+    /// If none are set, the default is to use the request's remote address field (as an ipStrategy).
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
@@ -907,10 +1182,13 @@ pub struct MiddlewareRateLimit {
     pub source_criterion: Option<MiddlewareRateLimitSourceCriterion>,
 }
 
-/// SourceCriterion defines what criterion is used to group requests as originating from a common source. If several strategies are defined at the same time, an error will be raised. If none are set, the default is to use the request's remote address field (as an ipStrategy).
+/// SourceCriterion defines what criterion is used to group requests as originating from a common source.
+/// If several strategies are defined at the same time, an error will be raised.
+/// If none are set, the default is to use the request's remote address field (as an ipStrategy).
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareRateLimitSourceCriterion {
-    /// IPStrategy holds the IP strategy configuration used by Traefik to determine the client IP. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ipallowlist/#ipstrategy
+    /// IPStrategy holds the IP strategy configuration used by Traefik to determine the client IP.
+    /// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ipallowlist/#ipstrategy
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
@@ -933,7 +1211,8 @@ pub struct MiddlewareRateLimitSourceCriterion {
     pub request_host: Option<bool>,
 }
 
-/// IPStrategy holds the IP strategy configuration used by Traefik to determine the client IP. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ipallowlist/#ipstrategy
+/// IPStrategy holds the IP strategy configuration used by Traefik to determine the client IP.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/ipallowlist/#ipstrategy
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareRateLimitSourceCriterionIpStrategy {
     /// Depth tells Traefik to use the X-Forwarded-For header and take the IP located at the depth position (starting from the right).
@@ -948,7 +1227,9 @@ pub struct MiddlewareRateLimitSourceCriterionIpStrategy {
     pub excluded_i_ps: Option<Vec<String>>,
 }
 
-/// RedirectRegex holds the redirect regex middleware configuration. This middleware redirects a request using regex matching and replacement. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/redirectregex/#regex
+/// RedirectRegex holds the redirect regex middleware configuration.
+/// This middleware redirects a request using regex matching and replacement.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/redirectregex/#regex
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareRedirectRegex {
     /// Permanent defines whether the redirection is permanent (301).
@@ -962,7 +1243,9 @@ pub struct MiddlewareRedirectRegex {
     pub replacement: Option<String>,
 }
 
-/// RedirectScheme holds the redirect scheme middleware configuration. This middleware redirects requests from a scheme/port to another. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/redirectscheme/
+/// RedirectScheme holds the redirect scheme middleware configuration.
+/// This middleware redirects requests from a scheme/port to another.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/redirectscheme/
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareRedirectScheme {
     /// Permanent defines whether the redirection is permanent (301).
@@ -976,7 +1259,9 @@ pub struct MiddlewareRedirectScheme {
     pub scheme: Option<String>,
 }
 
-/// ReplacePath holds the replace path middleware configuration. This middleware replaces the path of the request URL and store the original path in an X-Replaced-Path header. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/replacepath/
+/// ReplacePath holds the replace path middleware configuration.
+/// This middleware replaces the path of the request URL and store the original path in an X-Replaced-Path header.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/replacepath/
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareReplacePath {
     /// Path defines the path to use as replacement in the request URL.
@@ -984,7 +1269,9 @@ pub struct MiddlewareReplacePath {
     pub path: Option<String>,
 }
 
-/// ReplacePathRegex holds the replace path regex middleware configuration. This middleware replaces the path of a URL using regex matching and replacement. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/replacepathregex/
+/// ReplacePathRegex holds the replace path regex middleware configuration.
+/// This middleware replaces the path of a URL using regex matching and replacement.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/replacepathregex/
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareReplacePathRegex {
     /// Regex defines the regular expression used to match and capture the path from the request URL.
@@ -995,13 +1282,20 @@ pub struct MiddlewareReplacePathRegex {
     pub replacement: Option<String>,
 }
 
-/// Retry holds the retry middleware configuration. This middleware reissues requests a given number of times to a backend server if that server does not reply. As soon as the server answers, the middleware stops retrying, regardless of the response status. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/retry/
+/// Retry holds the retry middleware configuration.
+/// This middleware reissues requests a given number of times to a backend server if that server does not reply.
+/// As soon as the server answers, the middleware stops retrying, regardless of the response status.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/retry/
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareRetry {
     /// Attempts defines how many times the request should be retried.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub attempts: Option<i64>,
-    /// InitialInterval defines the first wait time in the exponential backoff series. The maximum interval is calculated as twice the initialInterval. If unspecified, requests will be retried immediately. The value of initialInterval should be provided in seconds or as a valid duration format, see https://pkg.go.dev/time#ParseDuration.
+    /// InitialInterval defines the first wait time in the exponential backoff series.
+    /// The maximum interval is calculated as twice the initialInterval.
+    /// If unspecified, requests will be retried immediately.
+    /// The value of initialInterval should be provided in seconds or as a valid duration format,
+    /// see https://pkg.go.dev/time#ParseDuration.
     #[serde(
         default,
         skip_serializing_if = "Option::is_none",
@@ -1010,15 +1304,28 @@ pub struct MiddlewareRetry {
     pub initial_interval: Option<IntOrString>,
 }
 
-/// StripPrefix holds the strip prefix middleware configuration. This middleware removes the specified prefixes from the URL path. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/stripprefix/
+/// StripPrefix holds the strip prefix middleware configuration.
+/// This middleware removes the specified prefixes from the URL path.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/stripprefix/
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareStripPrefix {
+    /// Deprecated: ForceSlash option is deprecated, please remove any usage of this option.
+    /// ForceSlash ensures that the resulting stripped path is not the empty string, by replacing it with / when necessary.
+    /// Default: true.
+    #[serde(
+        default,
+        skip_serializing_if = "Option::is_none",
+        rename = "forceSlash"
+    )]
+    pub force_slash: Option<bool>,
     /// Prefixes defines the prefixes to strip from the request URL.
     #[serde(default, skip_serializing_if = "Option::is_none")]
     pub prefixes: Option<Vec<String>>,
 }
 
-/// StripPrefixRegex holds the strip prefix regex middleware configuration. This middleware removes the matching prefixes from the URL path. More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/stripprefixregex/
+/// StripPrefixRegex holds the strip prefix regex middleware configuration.
+/// This middleware removes the matching prefixes from the URL path.
+/// More info: https://doc.traefik.io/traefik/v3.0/middlewares/http/stripprefixregex/
 #[derive(Serialize, Deserialize, Clone, Debug, Default, JsonSchema)]
 pub struct MiddlewareStripPrefixRegex {
     /// Regex defines the regular expression to match the path prefix from the request URL.
