+++ tembo-operator/src/traefik/mod.rs
@@ -1,8 +1,4 @@
 pub mod ingress_route_crd;
-pub mod ingress_route_crd_v3;
 pub mod ingress_route_tcp_crd;
-pub mod ingress_route_tcp_crd_v3;
 pub mod middleware_tcp_crd;
-pub mod middleware_tcp_crd_v3;
 pub mod middlewares_crd;
-pub mod middlewares_crd_v3;
