+++ tembo-operator/testdata/traefik-values.yaml
@@ -1,8 +1,7 @@
 image:
-  # This should exactly match the CRD version until
-  # Traefik v3 is released, because CRD is subject
-  # to breaking changes until it's released.
-  tag: v3.0.0-beta2
+  tag: v3.0.3-tembo.1
+  registry: quay.io/tembo
+  repository: traefik
 logs:
   general:
     level: DEBUG
@@ -18,7 +17,8 @@ additionalArguments:
   - "--api.debug=true"
 ports:
   postgresql:
-    expose: true
+    expose:
+      default: true
     port: 5432
     exposedPort: 5432
     nodePort: 32432
@@ -30,13 +30,15 @@ ports:
   #   nodePort: 32430
   #   protocol: TCP
   websecure:
-    expose: true
+    expose:
+      default: true
     port: 8443
     exposedPort: 8443
     nodePort: 32443
     protocol: TCP
   traefik:
-    expose: true
+    expose:
+      default: true
     port: 9000
     exposedPort: 9000
     nodePort: 32431
